pyproject.toml
[project]
name = "spring-into-haystack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "haystack-ai==2.12.2",
    "ipykernel>=6.29.5",
    "mcp-haystack",
]
